Gravity and movement error
Every time I would the player1 object it would move through the ground so I added a rigidbody and disabled gravity. It works but the player1 and player 2 object move up and down on the Y axis instead of the X axis. It should be playing like a 2d game but its playing like a 3d game